+++ src/bl4ckscor3/bot/bl4ckb0t/commands/channel/Info.java
@@ -8,6 +8,7 @@ import org.pircbotx.hooks.events.MessageEvent;
 import bl4ckscor3.bot.bl4ckb0t.Bot;
 import bl4ckscor3.bot.bl4ckb0t.Core;
 import bl4ckscor3.bot.bl4ckb0t.localization.L10N;
+import bl4ckscor3.bot.bl4ckb0t.util.TimeParser;
 import bl4ckscor3.bot.bl4ckb0t.util.Utilities;
 
 public class Info extends BaseCommand<MessageEvent<Bot>>
@@ -16,7 +17,7 @@ public class Info extends BaseCommand<MessageEvent<Bot>>
 	public void exe(MessageEvent<Bot> event, String[] args) throws Exception
 	{
 		Utilities.chanMsg(event, Colors.BOLD + "** " + L10N.getString("info.version", event).replace("#version", Colors.BOLD + Core.bot.getConfiguration().getVersion()) +
-				Colors.BOLD + " ** " + L10N.getString("info.uptime", event).replace("#uptime", Colors.BOLD + Utilities.millisToTimestamp(ManagementFactory.getRuntimeMXBean().getUptime())) + 
+				Colors.BOLD + " ** " + L10N.getString("info.uptime", event).replace("#uptime", Colors.BOLD + new TimeParser().lts(ManagementFactory.getRuntimeMXBean().getUptime(), "%s:%s:%s:%s")) + 
 				Colors.BOLD + " ** " + L10N.getString("info.author", event).replace("#author", Colors.BOLD + "bl4ckscor3") + Colors.BOLD + " **");
 	}