+++ src/bl4ckscor3/bot/bl4ckb0t/commands/channel/MCStatus.java
@@ -96,6 +96,6 @@ public class MCStatus extends BaseCommand<MessageEvent<Bot>>
 	@Override
 	public String[] getUsage(MessageEvent<Bot> event)
 	{
-		return new String[]{"-mcstatus || " + L10N.getString("mcsatus.explanation", event)};
+		return new String[]{"-mcstatus || " + L10N.getString("mcstatus.explanation", event)};
 	}
 }