+++ src/bl4ckscor3/bot/bl4ckb0t/commands/channel/Remind.java
@@ -1,52 +1,103 @@
 package bl4ckscor3.bot.bl4ckb0t.commands.channel;
 
-import java.util.concurrent.Executors;
-import java.util.concurrent.TimeUnit;
+import java.util.ArrayList;
 
 import org.pircbotx.hooks.events.MessageEvent;
 
 import bl4ckscor3.bot.bl4ckb0t.Bot;
 import bl4ckscor3.bot.bl4ckb0t.exception.IncorrectCommandExecutionException;
 import bl4ckscor3.bot.bl4ckb0t.localization.L10N;
+import bl4ckscor3.bot.bl4ckb0t.util.Reminder;
+import bl4ckscor3.bot.bl4ckb0t.util.TimeParser;
 import bl4ckscor3.bot.bl4ckb0t.util.Utilities;
 
 public class Remind extends BaseCommand<MessageEvent<Bot>>
-{	
+{
+	public static final ArrayList<Reminder> reminders = new ArrayList<Reminder>();
+
 	@Override
 	public void exe(MessageEvent<Bot> event, String[] args) throws Exception
 	{
-		if(args.length >= 3)
+		if(args.length == 2 && args[1].equals("list"))
 		{
-			String e = "";
-
-			for(int i = 2; i < args.length; i++)
+			if(reminders.size() > 0)
 			{
-				e += args[i] + " ";
+				String ids = "";
+				
+				for(Reminder r : reminders)
+				{
+					ids += "" + r.getId() + ", ";
+				}
+				
+				ids = ids.substring(0, ids.lastIndexOf(','));
+				Utilities.chanMsg(event, L10N.getString("remind.amount", event).replace("#amount", "" + reminders.size()).replace("#ids", ids));
 			}
-
-			String ev = e.trim();
-			long timeDue = 0;
+			else
+				Utilities.chanMsg(event, L10N.getString("remind.none", event));
 			
-			try
-			{
-				timeDue = new TimeParser().get(args[1]) * 1000L;
-			}
-			catch(NumberFormatException ex1)
-			{
-				throw new IncorrectCommandExecutionException(getMainAlias());
-			}
-			catch(IncorrectCommandExecutionException ex2)
-			{
-				return;
-			}
-			
-			Utilities.chanMsg(event, L10N.getString("remind.success", event));
-			Executors.newSingleThreadScheduledExecutor().schedule(() -> {
-				Utilities.chanMsg(event, L10N.getString("remind.reminder", event).replace("#user", event.getUser().getNick()).replace("#event", ev));
-			}, timeDue, TimeUnit.MILLISECONDS);
+			return;
 		}
-		else
-			throw new IncorrectCommandExecutionException(getMainAlias());
+		
+		try
+		{
+			int id = Integer.parseInt(args[1]); //if this doesn't fail we know that the user is probably asking for a specific id
+
+			if(!(id > Reminder.latestId))
+			{
+				for(Reminder r : reminders)
+				{
+					if(r.getId() == id)
+					{
+						if(args.length >= 3 && args[2].equals("stop"))
+						{
+							r.stop();
+							Utilities.chanMsg(event, L10N.getString("remind.stopped", event).replace("#id", "" + r.getId()));
+							return;
+						}
+						
+						Utilities.chanMsg(event, L10N.getString("remind.timeLeft", event).replace("#event", r.getEvent()).replace("#timeLeft", new TimeParser(getMainAlias()).lts(r.getRemainingTime(), "%sd%sh%sm%ss")));
+						return;
+					}
+				}
+			}
+			
+			Utilities.chanMsg(event, L10N.getString("remind.noId", event));
+		}
+		catch(NumberFormatException ex)
+		{
+			if(args.length >= 3)
+			{
+				String e = "";
+
+				for(int i = 2; i < args.length; i++)
+				{
+					e += args[i] + " ";
+				}
+
+				String ev = e.trim();
+				long timeDue = 0;
+
+				try
+				{
+					timeDue = new TimeParser(getMainAlias()).stl(args[1]) * 1000L;
+				}
+				catch(NumberFormatException ex1)
+				{
+					throw new IncorrectCommandExecutionException(getMainAlias());
+				}
+				catch(IncorrectCommandExecutionException ex2)
+				{
+					return;
+				}
+
+				Reminder reminder = new Reminder(event, ev, timeDue);
+				
+				Utilities.chanMsg(event, L10N.getString("remind.success", event).replace("#id", "" + reminder.getId()));
+				reminders.add(reminder);
+			}
+			else
+				throw new IncorrectCommandExecutionException(getMainAlias());
+		}
 	}
 
 	@Override
@@ -64,116 +115,9 @@ public class Remind extends BaseCommand<MessageEvent<Bot>>
 	@Override
 	public String[] getUsage(MessageEvent<Bot> event)
 	{
-		return new String[]{"-remind 1d1h1m1s abc|| " + L10N.getString("remind.explanation", event)};
-	}
-
-	private class TimeParser
-	{
-		public long get(String s) throws IncorrectCommandExecutionException
-		{
-			if(s.contains("d"))
-				return d(s);
-			else if(s.contains("h"))
-				return h(s);
-			else if(s.contains("m"))
-				return m(s);
-			else if(s.contains("s"))
-				return s(s);
-			else
-				throw new IncorrectCommandExecutionException(getMainAlias());
-		}
-
-		private long d(String s)
-		{
-			long value = Long.parseLong(s.split("d")[0]);
-
-			value *= 24 * 60 * 60;
-
-			if(s.contains("h"))
-				return value + h(s);
-			else if(s.contains("m"))
-				return value + m(s);
-			else if(s.contains("s"))
-				return value + s(s);
-			else
-				return value;
-		}
-
-		private long h(String s)
-		{
-			long value = 0;
-
-			if(s.contains("d"))
-				value = Long.parseLong(s.split("d")[1].split("h")[0]);
-			else
-				value = Long.parseLong(s.split("h")[0]);
-
-			value *= 60 * 60;
-
-			if(s.contains("m"))
-				return value + m(s);
-			else if(s.contains("s"))
-				return value + s(s);
-			else
-				return value;
-		}
-
-		private long m(String s)
-		{
-			long value = 0;
-
-			if(s.contains("d"))
-			{
-				if(s.contains("h"))
-					value = Long.parseLong(s.split("d")[1].split("h")[1].split("m")[0]);
-				else
-					value = Long.parseLong(s.split("d")[1].split("m")[0]);
-			}
-			else if(s.contains("h"))
-				value = Long.parseLong(s.split("h")[1].split("m")[0]);
-			else
-				value = Long.parseLong(s.split("m")[0]);
-
-			value *= 60;
-
-			if(s.contains("s"))
-				return value + s(s);
-			else
-				return value;
-		}
-
-		private long s(String s)
-		{
-			long value = 0;
-
-			if(s.contains("d"))
-			{
-				if(s.contains("h"))
-				{
-					if(s.contains("m"))
-						value = Long.parseLong(s.split("d")[1].split("h")[1].split("m")[1].split("s")[0]);
-					else
-						value = Long.parseLong(s.split("d")[1].split("h")[1].split("s")[0]);
-				}
-				else
-				{
-					if(s.contains("m"))
-						value = Long.parseLong(s.split("d")[1].split("m")[1].split("s")[0]);
-					else
-						value = Long.parseLong(s.split("d")[1].split("s")[0]);
-				}
-			}
-			else if(s.contains("h"))
-			{
-				if(s.contains("m"))
-					value = Long.parseLong(s.split("h")[1].split("m")[1].split("s")[0]);
-				else
-					value = Long.parseLong(s.split("h")[1].split("s")[0]);
-			}
-			else if(s.contains("m"))
-				value = Long.parseLong(s.split("m")[1].split("s")[0]);
-
-			return value;
-		}
+		return new String[]{
+				"-remind 1d1h1m1s abc || " + L10N.getString("remind.explanation.1", event),
+				"-remind 5 || " + L10N.getString("remind.explanation.2", event)
+		};
 	}
 }