+++ src/bl4ckscor3/bot/bl4ckb0t/util/Utilities.java
@@ -301,20 +301,6 @@ public class Utilities
 		return s.startsWith("1") || s.startsWith("2") || s.startsWith("3") || s.startsWith("4") || s.startsWith("5") || s.startsWith("6") || s.startsWith("7") || s.startsWith("8") || s.startsWith("9") || s.startsWith("0");
 	}
 
-	public static String millisToTimestamp(long millis)
-	{
-		long seconds = millis / 1000;
-		long minutes = seconds / 60;
-		long hours = minutes / 60;
-		long days = hours / 24;
-
-		seconds -= 60 * minutes;
-		minutes -= 60 * hours;
-		hours -= 24 * days;
-		
-		return String.format("%s:%s:%s:%s", (days < 10 ? "0" + days : days), (hours < 10 ? "0" + hours : hours), (minutes < 10 ? "0" + minutes : minutes), (seconds < 10 ? "0" + seconds : seconds));
-	}
-
 	/**
 	 * Checks wether a user is ignored
 	 * @param nick The user to check